fix(demos): Pixel Prio: fix rendering, memo color calculation, refactor code
